KAFKA-18329; [1/3] Delete old group coordinator (KIP-848) (#19243)
This patch is the first of a series of patches to remove the old group
coordinator. With the release of Apache Kafka 4.0, the so-called new
group coordinator is the default and only option available now.

This patch update the system tests to not run with the old group
coordinator. It also removed the ability to use the old group
coordinator.

Reviewers: Lianet Magrans <lmagrans@confluent.io>