KAFKA-15931: Cancel RemoteLogReader gracefully (#19197)
Reverts commit
https://github.com/jeqo/kafka/commit/2723dbf3a007af0a1cafc17ed019aabd46eb6881
and
https://github.com/jeqo/kafka/commit/269e8892ad26904c82a3bff7b7d7a81c47619e4a.

Instead of reopening the transaction index, it cancels the
RemoteFetchTask without interrupting it--avoiding to close the
TransactionIndex channel.

This will lead to complete the execution of the remote fetch but ignoring
the results. Given that this is considered a rare case, we could live
with this. If it becomes a performance issue, it could be optimized.

Reviewers: Jun Rao <junrao@gmail.com>