KAFKA-18980 OffsetMetadataManager#cleanupExpiredOffsets should record the number of records rather than topic partitions (#19207)
jira: https://issues.apache.org/jira/browse/KAFKA-18980

We should use the number of `records` rather than topic partitions,but
this is not a bug as the number of `records` should be equal to the
number of topic partitions.

Also, we previously used `expiredPartitions` for logging, but now it is
not being used anymore, so we can remove it.

Reviewers: TengYao Chi <kitingiao@gmail.com>, Chia-Ping Tsai
<chia7712@gmail.com>