KAFKA-19016: Updated the retention behaviour of share groups to retain them forever (#19343)
This PR makes the required changes for retaining a share group forever
post creation. A new field `shouldExpire` is added to all types
implementing the interface `Group`. It returns false only in case of
`ShareGroup` and true otherwise. A check is also added in
`cleanupGroupMetadata` in `GroupCoordinatorShard` and it proceeds with
group expiration only if `shouldExpire` returns true

Reviewers: Andrew Schofield <aschofield@confluent.io>, PoAn Yang <payang@apache.org>